Finished Project!
I figured out the cause of the infinite loop. I was using global queues and forgetting to empty them. I localized them so they start as queues with just the start node in instead of nodes from the previous bfs call. Also, changed the function name from dfs to bfs.